chore (refs T22614): move DpContextualHelp to client/js.
