+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/Shared/SimilarStatementSubmitters/SimilarStatementSubmitters.vue
@@ -156,7 +156,7 @@
 
 <script>
 import { checkResponse, dpApi } from '@DemosPlanCoreBundle/plugins/DpApi'
-import DpContextualHelp from '@DemosPlanCoreBundle/components/DpContextualHelp'
+import DpContextualHelp from '@DpJs/components/core/DpContextualHelp'
 import DpEditableList from '@DpJs/components/core/DpEditableList'
 import { DpInput } from 'demosplan-ui/components'
 import dpValidateMixin from '@DpJs/lib/validation/dpValidateMixin'
